# src/app/models/carrera.model.ts
export class Carrera {
    idCarrera: number;
    nombreCarrera: string
    codigoCarrera: string;
    idInstituto: number;
    siglas: string;
    constructor() {
        this.idCarrera=0;
        this.nombreCarrera='Diseño Grafico';
        this.codigoCarrera= '01';
        this.idInstituto=0;
        this.siglas='DG';
    }
};