# src/app/models/instituto.model.ts
export class Instituto{
    idInstituto:number;
    nombreInstituto: string;
    codigoInstituto: string;
    constructor() {
        this.idInstituto= 0;
        this.nombreInstituto = 'Instituto de sistemas';
        this.codigoInstituto = '099';
    }
}